CN: first attempt at dealing with C type-names in the new CN parser
This does not duplicate the C grammar, instead when the CN lexer
sees matching `<` ... `>` it attempts to parse a C type-name between
the angles (using a new entry point to the C parser) and produces
a single `LT_CTYPE_GT` token for the CN parser.

This may fail because the opening `<` may in fact be a relational
operator. In this case the CN lexer recovers by re-lexing from
the top.

This first attempt avoids changing the existing CN syntax. If we
instead change it to have dedicated delimiters around C type-names,
we won't need to do the re-lexing recovery.

Most of the CN ci tests now work (the failures are caused by the current
lack of pretty error messages in the CN parser), but this needs proper
testing...