[CN] preserving `Nothing` elements in `AilSdeclaration` during desugaring
These were previously filtered out to prevent the elaboration from producing
stores of unspecified values (which CN does not have).

This was an issue from CN-exec.

Now the Ail generated by the desugaring is the same as in base Cerberus,
and it is the elaboration does something special for CN.