Lexing: staging the creation/initialisation of C/CN lexers
The C/CN lexer uses a reference to remember the kind of the last seen token.
As a result, if it used for multiple parsing sessions this state
must be reset. Forgetting to do so caused issue #382.

This commit stages the creation of lexers to make it harder
to forget using a fresh lexer when starting a new C or CN parsing session.

(based to a diff by @dc-mak)