CN: Add support for bitwise subtraction
Fixes https://github.com/rems-project/cerberus/issues/255

Though the test looks like it is for pointer difference, the actual
issue is that the elaboration has `array_shift( a_533, signed int,   0 -
a_534)`, and it is the `0 - a534` which was causing the error.

In principle we can/should have pure subtraction at integer and real
types as well, but this can be generalised later as and when the need
arises.