CN: Run docker build nightly (#489)
The docker build is slow, especially on Arm since it has to run on QEMU
(OCaml doesn't support cross compilation well yet) and because of this,
and the `cancel-in-progress: true`, the CI becomes less helpful (pushing
new commits cancels the previous job, which is fine, but then that shows
up as failed on the CI history, which is unhelpful).

This commit changes it to run the build daily at 18:30 UTC and disables
cancel-in-progress.