CN: Sundry fixes to solver file
Struct and data fields are now suffixed with an underscore, making them
more readable (and more consistent).

"XXX" is replaced with TODO, FIXME, or "ISD" as approriately.

"intptr_cast" was renamed to the more accurate "uintptr_cast" (no
change in semantics, just a variable renaming).

A typo "done_strcuts" was changed to "done_structs"

A variable "here" which did not point to the OCaml function was renamed
to "loc" (no change in semantics, merely renaming).

Two constructors "BWOrNoSMT" and "XORNoSMT" were given simpler, more
readable and consistent names "BW_Or" and "BW_Xor".