CN: Add test for Z3 let var calc
This is a test which happens to trigger Z3 (4.13.0) to produce a
particular kind of let-binding in a response, thus exposing a bug in the
free variable calculation. This bug was fixed in the below commit, but
without a repro/test (the bug was discovered on an incomplete feature
branch).
https://github.com/rems-project/cerberus/commit/4e7fcb9f29eb9f7bbc49c092c2850f7da9e8dd66