CN VIP: Add allocation history constraints
This commit enables (under the `--vip` flag) the Create memory action in
CN adding a constraint to the `allocs` allocation history variable to
map the newly allocated pointer's allocation ID to the appropriate base
and length.

This change is underneath a flag because it is not backwards compatiable
in the presence of more than one create (of a different adress or size)
because in the solver, all allocation IDs are mapped to unit, meaning
adding the second constraint makes the context inconsistent and
terminates type checking early.

As a result, the test suite is now run with the `--vip` flag enabled (so
that new features can be tested).  Note that the tutorial remains
running without the `--vip` flags (and should not yet be using any of
the new featuers), and so any backwards compatibility issues will be
caught there.