CN VIP: Generalise ptr diff support
This commit generalises the support for pointer subtraction by
(a) working around an pointer representation quirk (b) working around
another Z3 bug by extending the simplifier (c) removing an unnecessary
constraint check from the live allocation check.

Sometimes we may wish to write `Owned(array_shift(p, x))`, and because
of the solver represenation of array_shifting, the solver cannot
conclude that `has_alloc_id(array_shift(p,x))` implies `has_alloc_id(p)`
(e.g. p == NULL and has_alloc_id(default<Loc>) are possible). So,
the smart constructor for `has_alloc_id` does this for us syntactically.
If this happens semantically, then the user will just need to fix their
specs - this is mostly just so that deriving non-null constraints from
Owneds still works as one would expect.

The Z3 bug is basically not being able to figure out that `(p + 2) - 2
== p` for a pointer p (so the + is actually an array_shift), so the
simplifier takes care of that.

There is no need to do a bounds check in the resource loop since that is
always done in the check.ml as part of the typing rule - the previous
code was too restrictive (you can validly learn the allocation is live
even in the requested pointer doesn't fall in the footprint of an Owned
in the context).