CN: address https://github.com/rems-project/cerberus/issues/399
This refines the resource tracking in the wellformedness checks of CN
specifications, fixing previous a previous spurious error about
"inconsistent assumptions".

The issue is the following: CN's internal representation of C
functions is:
- a collection of labeled code blocks and
- a designated function body, which may jump to any of the blocks

The logical arguments (including resources) to a function are in the
scope of the labels, because their specifications may refer to
variables introduced in the precondition, but the *resources* should
not be because they have to be explicitly passed into the labels at
the callsite.

CN's main typechecking logic, in check.ml, already correctly handled
the scoping of resources within labels in this regard, but until
https://github.com/rems-project/cerberus/pull/385 the wellTyped logic
had no reason to ever "forget" resources within a function.