CN VIP: Add bounds checks for shifting
This commit implements the VIP typing rules for member and array
shifting. In particular, pure shifts do not require the allocation to be live,
but the ISO version (PtrArrayShift) does. However, that AST construct is
only produced if the appropriate switch is passed to Cerberus, and so
conditioning on that is not required within the CN/checker code.