CN: Fix #598 (solver crash) reduce mutability
PR #251 added support for a new solver backend, but also introduced a
bug which, with `--vip` enabled, caused the solver to crash.

The root cause was that calls to `provable` were interfering with
`model_evaluator` because of unnecessary shared mutability.

This can be fixed by calling `Int_Table.copy` in `copy_solver_frame`,
but instead this removes the two-layer hashtable and replaces it with a
simpler pair-lookup map.