CN VIP: Add conversion to/from bytes for Blocks
This commit adds conversion to/from bytes for Blocks. In doing so,
it restricts the grammar for to_bytes/from_bytes to only accept
predicates, because we need the C type information when converting
from bytes, and can derive the bytes resource to be requested in the
same way as we do for to_bytes.

This is easy because the hard parts about provenances in bytes and
constraining values appropriately only show up with Owned resources.