CN VIP: Update the succ tests to use byte rep conv
Because of the lack of roundtrip support (not just via u/intptr_t, but
also via byte representations) this commit adds a `#ifdef NO_ROUND_TRIP`
line to many tests to compensate by using a copy_alloc_id (though this
is possible in these tests, it won't always be possible in general).

All the added lines are of the form `p = copy_alloc_id((uintptr_t)p, q)`
which means that the prior code must get the address correct for only
the allocation ID to be fixed up.

Also fixes a pointer-to-integer (index term) casting bug and changes the
base type to explicitly be unsigned char (char being signed or unsigned
is implementation-defined, could have just as well picked signed char).