CN VIP: Add support for memcpy & proxy
This commit hooks into an existing place for assigning function types to
proxy functions. It assumes that memcpy's type is expressible in CN,
which is true-ish for now, but may not remain so if we want to support
round-trip for pointers without provenance in byte values, and thus need
a "special type" or more arbitrary set of operations on the typing
context to support.

It also removes some legacy/mostly unused code around partial
evaluation.