CN VIP: optimise shifts
For some reason, checking the resource context is faster than doing a
bounds check. Since having an Owned in the context is sufficient (but
not necessary) to proving a pointer is in bounds, and often a pointer
will be constructed just before it is dereferenced, then it suffices to
check the context for the exact pointer as a common case, and fall back
to the slow path for the rest.