CN: Assert max tuple size
In the buddy allocator, tuples can end up being larger than 8, and so
the error comes up as a solver rather than a more helpful arity one.
This commit inserts a quick fix to make it fail at a better location. It
does not come with a test because the buddy allocator is too large and
slow to cut down, and I don't know how to coax Cerberus into generating
tuples that big.