Track location in pre-processed file, as well references to source.
We do this by introducing a new module `Cerb_location.position`, which
replaces `Lexing.position` as the base type to use for positions.
Values of the new position type keepl the `Lexing.position`
corresponding to the actual location in the preprocessed file,
as well as the file name and line in the original source assocaited with
the location.