[CN-Seq-Test] Add `cn seq_test` (#855)
USAGE:

1. cd to the directory containing the file you want to test.
2. Suppose your file is called test.c.
Run the following command to generate a sequence of 100 random calls. "cn seq_test test.c --output-dir="test"
The output-dir parameter is required. The call above places the resulting files in the test directory.
If a nonexistent directory is passed in, the directory will be automatically created and the test files placed in there.
There is an optional --num-samples parameter that specifies the number of calls to generate.
Some other useful options are --max-backtrack-attempts and --max-resets.
--max-backtrack-attempts sets maximum the number of times to attempt generating a call.
--max-resets sets the maximum number of times to clear the generation context (which records the previous calls
for reuse in later calls).
3. After running "cn seq_test..." there will be a run_tests.sh script. Running the script will link all the object files,
create the executable as "tests.out" and run the executable.