Add typedef annotation for Byte
This commit adds support for translating any `typedef name unsigned
char;` as a byte with a name chosen by the user, by prefixing it with
`[[cerb::byte]]`.

It also adds a constraint failure if the attribute is used on a typedef
for a type which is not an unsigned char.