Make bytes casts pure memops instead
As per review feedback, this commit changes the type of byte casts
to be pure memops rather than effectful ones (since they don't use any
ghost state). The UB associated with creating a byte from an integer
is made explicit in the Core elaboration.

The reason for keeping it as a memop is that in CN the basetypes and
representation of bytes is distinct to that of integers, unlike in Core
where they happily coincide. Thus conversions between them need to be
explicitly marked as such.